Django blog initialize
The theme is based on the Flex, it's the theme I used in my pelican
blog, so I copy it here and implement the most feature is supports.

And this is a repo that I want to use to try some integration tools
like Travis on github, so it's also work as an experiment lab.

Reference:
- Flex: https://github.com/alexandrevicenzi/flex